feat: Add Docker Compose configuration and implement point table functionality
