+++ frontend/partie.html
@@ -13,7 +13,8 @@
   <script type="importmap">
     {
       "imports": {
-        "partieberechnungen": "./partieberechnungen.js"
+        "partieberechnungen": "./partieberechnungen.js",
+        "partietabelle": "./partietabelle.js"
       }
     }
   </script>
